Created the Lexical Order GUID based on the ULID specification

+++ src/main/java/com/github/f4b6a3/uuid/clockseq/ClockSequenceStrategy.java
@@ -1,5 +1,5 @@
 package com.github.f4b6a3.uuid.clockseq;
 
 public interface ClockSequenceStrategy {
-	int getClockSequence(long timestamp, long nodeIdentifier);
+	long getClockSequence(long timestamp, long nodeIdentifier);
 }
