+++ src/main/java/com/github/f4b6a3/uuid/clockseq/CombClockSequenceStrategy.java
@@ -16,7 +16,7 @@ public class CombClockSequenceStrategy extends AbstractSequence implements Clock
 	}
 
 	@Override
-	public int getClockSequence(final long timestamp, final long nodeIdentifier) {
+	public long getClockSequence(final long timestamp, final long nodeIdentifier) {
 		if (nodeIdentifier != this.nodeIdentifier) {
 			this.nodeIdentifier = nodeIdentifier;
 			this.reset();
