+++ src/main/java/com/github/f4b6a3/uuid/clockseq/FixedClockSequenceStrategy.java
@@ -9,7 +9,7 @@ public class FixedClockSequenceStrategy implements ClockSequenceStrategy {
 	}
 
 	@Override
-	public int getClockSequence(long timestamp, long nodeIdentifier) {
+	public long getClockSequence(long timestamp, long nodeIdentifier) {
 		return this.clockSequence;
 	}
 }
