+++ src/main/java/com/github/f4b6a3/uuid/state/AbstractUuidState.java
@@ -3,7 +3,7 @@ package com.github.f4b6a3.uuid.state;
 public abstract class AbstractUuidState {
 
 	protected long timestamp = 0;
-	protected int clockSequence = 0;
+	protected long clockSequence = 0;
 	protected long nodeIdentifier = 0;
 
 	public AbstractUuidState() {
@@ -17,11 +17,11 @@ public abstract class AbstractUuidState {
 		this.timestamp = timestamp;
 	}
 
-	public int getClockSequence() {
+	public long getClockSequence() {
 		return clockSequence;
 	}
 
-	public void setClockSequence(int clockSequence) {
+	public void setClockSequence(long clockSequence) {
 		this.clockSequence = clockSequence;
 	}
